move native list-param SQL generation from backends into frontends
Before this change each backend branched on its Engine/Target enum to
produce dialect-specific SQL for native list params (= ANY($N),
json_each, JSON_TABLE). This violated the N+M goal: backends contained
SQL knowledge that belongs to frontends.

Changes:

- Add `native_list_sql: Option<String>` to `Parameter` in the IR.
  Frontends populate this field for every list parameter using a new
  `ResolverConfig::native_list_sql` function pointer. The SQL uses $N
  placeholders so backends apply their standard rewriting uniformly.

- Each dialect frontend now owns its own native-list SQL production:
  - postgres: `replace_list_in_clause` → `= ANY($N)`
  - sqlite:   `replace_list_in_clause` → `IN (SELECT value FROM json_each($N))`
  - mysql:    `replace_list_in_clause` → `IN (SELECT value FROM JSON_TABLE(...))`

- All backends simplified: they read `lp.native_list_sql` from the IR
  and apply their standard placeholder rewriting. No backend now
  contains dialect-specific SQL string construction.

- Remove the last dialect branch in `rust_type`: SQLite's NUMERIC/DECIMAL
  types now map to `SqlType::Double` in the SQLite typemap (REAL affinity),
  so the Rust backend no longer needs a target-dependent branch for Decimal.
  Golden files and tests updated accordingly.

- Add `Parameter::with_native_list_sql` builder for use in tests that
  construct Query objects directly (bypassing the frontend).

+++ src/frontend/sqlite/schema.rs
@@ -215,6 +215,7 @@ mod tests {
         assert_eq!(col("t").sql_type, SqlType::Text);
         assert_eq!(col("r").sql_type, SqlType::Real);
         assert_eq!(col("b").sql_type, SqlType::Bytes);
-        assert_eq!(col("n").sql_type, SqlType::Decimal);
+        // SQLite has no fixed-point type; NUMERIC maps to Double (REAL affinity).
+        assert_eq!(col("n").sql_type, SqlType::Double);
     }
 }
